refactor(transfermarkt): extract live-match parsing to live_parser.py
Reduces parser.py from 609 to 232 LOC. LiveMatchParser(ParseHelpers)
owns all live-match HTML parsing; TransfermarktParser delegates via a
one-line parse_live_matches(). Also merges the two near-identical
_extract_competition_info_from_* methods into one
_extract_competition_info(header, *, flag_search_in) call.
